plus aue le test et le bonus

+++ Projets TC/Libft/ft_calloc.c
@@ -6,16 +6,16 @@
 /*   By: flee <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/08 10:26:08 by flee              #+#    #+#             */
-/*   Updated: 2021/06/08 14:15:27 by flee             ###   ########.fr       */
+/*   Updated: 2021/06/10 12:34:57 by flee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
 void	*ft_calloc(size_t elementCount, size_t elementSize)
 {
-	char	*memory;
-	int		index;
+	char			*memory;
+	unsigned long	index;
 
 	memory = malloc(elementSize * elementCount);
 	index = 0;
