+++ Projets TC/Libft/ft_memchr.c
@@ -6,17 +6,17 @@
 /*   By: flee <marvin@42.fr>                        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/07 10:49:46 by flee              #+#    #+#             */
-/*   Updated: 2021/06/08 13:53:19 by flee             ###   ########.fr       */
+/*   Updated: 2021/06/10 12:35:40 by flee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
 void	*ft_memchr(const void *memoryBlock, int searchedChar, size_t size)
 {
-	char	*memory;
-	char	needle;
-	int		index;
+	char			*memory;
+	char			needle;
+	unsigned long	index;
 
 	memory = (char *)memoryBlock;
 	needle = (char)searchedChar;
