+++ Projets TC/Libft/ft_memcpy.c
@@ -6,17 +6,17 @@
 /*   By: flee <flee@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/04 15:43:48 by flee              #+#    #+#             */
-/*   Updated: 2021/06/08 13:40:35 by flee             ###   ########.fr       */
+/*   Updated: 2021/06/10 12:36:29 by flee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	char	*b;
-	char	*a;
-	int		index;
+	char				*b;
+	char				*a;
+	unsigned long		index;
 
 	b = (char *)dest;
 	a = (char *)src;
