+++ Projets TC/Libft/ft_strnstr.c
@@ -10,12 +10,12 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	index;
-	int	j;
+	unsigned long	index;
+	unsigned long	j;
 
 	index = 0;
 	if (needle[0] == '\0')
